Validation of application forms.
In business sector job applications, the age of the applicant is
calculated and set if they are going to be over eighteen when the job
starts.
The date of birth is saved in both types of applications. 
Fixed the SuccessView so that it works for when you've applied for a
job.

Etc.